+++ vibeio/src/op/readiness.rs
@@ -7,7 +7,6 @@ use crate::op::Op;
 
 pub struct ReadinessOp<'a> {
     handle: &'a InnerRawHandle,
-    ready: bool,
     is_writable: bool,
 }
 
@@ -16,7 +15,6 @@ impl<'a> ReadinessOp<'a> {
     pub fn new_readable(handle: &'a InnerRawHandle) -> Self {
         Self {
             handle,
-            ready: false,
             is_writable: false,
         }
     }
@@ -25,7 +23,6 @@ impl<'a> ReadinessOp<'a> {
     pub fn new_writable(handle: &'a InnerRawHandle) -> Self {
         Self {
             handle,
-            ready: false,
             is_writable: true,
         }
     }
@@ -40,21 +37,15 @@ impl Op for ReadinessOp<'_> {
         cx: &mut Context<'_>,
         driver: &AnyDriver,
     ) -> Poll<io::Result<Self::Output>> {
-        // For now, use "ready" state. This might cause "readiness" on spurious wakeups, but it should work just fine.
-        if self.ready {
-            Poll::Ready(Ok(()))
-        } else {
-            self.ready = true;
-            driver.submit_poll(
-                self.handle,
-                cx.waker().clone(),
-                if self.is_writable {
-                    mio::Interest::WRITABLE
-                } else {
-                    mio::Interest::READABLE
-                },
-            )?;
-            Poll::Pending
-        }
+        driver.submit_poll(
+            self.handle,
+            cx.waker().clone(),
+            if self.is_writable {
+                mio::Interest::WRITABLE
+            } else {
+                mio::Interest::READABLE
+            },
+        )?;
+        Poll::Pending
     }
 }