+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Added `AsyncReadPoll` and `AsyncWritePoll` traits for poll-based async read/write readiness interfaces.
 - Added `PollUdpSocket` struct for poll-based UDP socket readiness interfaces.
+- Added `time` Cargo feature to `vibeio-hyper` that can be optionally disabled.
+- Added `try_io_readable` and `try_io_writable` methods to `PollUdpSocket`, `PollTcpStream`, and `PollUnixStream`.
+- Initial version of `vibeio-quinn` crate for poll-based QUIC socket readiness interfaces.
+- `vibeio-hyper` now no longer depends on all the Cargo features of `vibeio`.
 
 ## `vibeio` 0.2.13